Refactor `HashMapper` to depend on a base class
`MetadataMapper` is introduced as an abstract/Null mapper class, from which
others inherit. It provides a base `#fields` implementation that returns an
empty `Array`, as well as `#field`, the `#metadata` accessor, `#field?`, and a
base `#method_missing` that can be customized by providing `#map_field`.
