+++ lib/darlingtonia.rb
@@ -6,6 +6,7 @@ require 'active_fedora'
 # Bulk object import for Hyrax
 module Darlingtonia
   require 'darlingtonia/version'
+  require 'darlingtonia/metadata_mapper'
   require 'darlingtonia/hash_mapper'
 
   require 'darlingtonia/importer'
